% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harness_utils.R
\name{confusion_raw}
\alias{confusion_raw}
\title{Build raw (integer) confusion matrix from y, y_pred.}
\usage{
confusion_raw(y, y_pred, C)
}
\description{
Build raw (integer) confusion matrix from y, y_pred.
}
