% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harness_utils.R
\name{confusion_weighted}
\alias{confusion_weighted}
\title{Build weighted confusion matrix from y, y_pred, w.}
\usage{
confusion_weighted(y_true, y_hat, w, C = 3L)
}
\description{
Build weighted confusion matrix from y, y_pred, w.
}
