% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unioda_core.R
\name{oda_apply_primary_secondary}
\alias{oda_apply_primary_secondary}
\title{Apply primary/secondary tie-breaking to candidates}
\usage{
oda_apply_primary_secondary(cand_df, primary, secondary, y, w, preds_list)
}
\description{
Apply primary/secondary tie-breaking to candidates
}
\keyword{internal}
