% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_apply_priors_multiclass}
\alias{oda_apply_priors_multiclass}
\title{Class-normalise weights so each class sums to 1 (PRIORS ON).}
\usage{
oda_apply_priors_multiclass(y, w, priors_on = TRUE)
}
\description{
Class-normalise weights so each class sums to 1 (PRIORS ON).
}
