% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unioda_core.R
\name{oda_clean_xy}
\alias{oda_clean_xy}
\title{Drop observations missing on x, y, or w for this attribute only.}
\usage{
oda_clean_xy(x, y, w = NULL, miss_codes = NULL)
}
\description{
Drop observations missing on x, y, or w for this attribute only.
}
