% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_enforce_weighting_policy}
\alias{oda_enforce_weighting_policy}
\title{Enforce weighting policy for multiclass ODA}
\usage{
oda_enforce_weighting_policy(
  attr_type_res,
  priors_on,
  loo,
  w_case,
  reason_prefix = ""
)
}
\description{
Enforce weighting policy for multiclass ODA
}
\keyword{internal}
