% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_is_nondegenerate_labels}
\alias{oda_is_nondegenerate_labels}
\title{Non-degeneracy check: predicted labels must cover all C classes.}
\usage{
oda_is_nondegenerate_labels(pred_labels, C)
}
\description{
Non-degeneracy check: predicted labels must cover all C classes.
}
