% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_loo_multiclass}
\alias{oda_loo_multiclass}
\title{Leave-one-out for categorical multiclass ODA}
\usage{
oda_loo_multiclass(
  x,
  y,
  w0,
  attr_type,
  priors_on_eff,
  weighted_requested,
  degen,
  K_segments,
  miss_codes = NULL
)
}
\description{
Leave-one-out for categorical multiclass ODA
}
