% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unioda_core.R
\name{oda_make_blocks_ordered}
\alias{oda_make_blocks_ordered}
\title{Collapse sorted x into unique-value blocks; return cumulative class counts.}
\usage{
oda_make_blocks_ordered(x, y, w)
}
\description{
Collapse sorted x into unique-value blocks; return cumulative class counts.
}
