% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_make_blocks_ordered_multiclass}
\alias{oda_make_blocks_ordered_multiclass}
\title{Build per-class count matrix over unique-x blocks (objective or raw space).}
\usage{
oda_make_blocks_ordered_multiclass(x, y, w)
}
\description{
Build per-class count matrix over unique-x blocks (objective or raw space).
}
