% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_make_global_blocks_ordered}
\alias{oda_make_global_blocks_ordered}
\title{Build global block grid on full x (for LOO fixed-grid mode).}
\usage{
oda_make_global_blocks_ordered(x)
}
\description{
Build global block grid on full x (for LOO fixed-grid mode).
}
