% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_mc_p_value_multiclass}
\alias{oda_mc_p_value_multiclass}
\title{Monte Carlo p-value for multiclass ODA}
\usage{
oda_mc_p_value_multiclass(
  x,
  y,
  w,
  attr_type,
  priors_on,
  degen,
  K_segments,
  mc_iter = 25000L,
  mc_target = 0.05,
  mc_stop = 99.9,
  mc_stopup = NA,
  mc_adjust = FALSE,
  seed = NULL,
  observed_mean_pac,
  direction = "off",
  direction_map = NULL
)
}
\description{
Monte Carlo p-value for multiclass ODA
}
\keyword{internal}
