\name{oda_mean_pac}
\alias{oda_mean_pac}
\title{Mean PAC from sensitivity and specificity}
\description{Compute mean Percentage Accuracy in Classification.}
\usage{oda_mean_pac(sens, spec)}
\arguments{
  \item{sens}{Sensitivity (proportion [0, 1]).}
  \item{spec}{Specificity (proportion [0, 1]).}
}
\value{Mean PAC as a proportion [0, 1].}
