% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioda_core.R
\name{oda_metrics_candidate}
\alias{oda_metrics_candidate}
\title{Compute PAC, SAMPLEREP, and other metrics for a partition}
\usage{
oda_metrics_candidate(
  NP_obj,
  NP_raw,
  NA_raw,
  x_rep,
  cuts_idx,
  priors_on_eff = TRUE
)
}
\description{
Compute PAC, SAMPLEREP, and other metrics for a partition
}
\keyword{internal}
