% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unioda_core.R
\name{oda_resolve_attr_type}
\alias{oda_resolve_attr_type}
\title{Resolve attribute type from data if "auto".}
\usage{
oda_resolve_attr_type(
  x,
  attr_type = c("auto", "ordered", "categorical", "binary")
)
}
\description{
Resolve attribute type from data if "auto".
}
