% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unioda_core.R
\name{oda_rule_side}
\alias{oda_rule_side}
\title{Assign each observation to the left (0) or right (1) side of a rule.}
\usage{
oda_rule_side(x, rule)
}
\description{
Assign each observation to the left (0) or right (1) side of a rule.
}
