% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_auto_plan.R
\name{print.auto_sda_plan}
\alias{print.auto_sda_plan}
\title{Print an auto_sda_plan object}
\usage{
\method{print}{auto_sda_plan}(x, ...)
}
\arguments{
\item{x}{An \code{auto_sda_plan} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}. Called primarily for its side effect of
  printing a human-readable summary of the SDA plan to the console.
}
\description{
Print an auto_sda_plan object
}
