% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta_ort.R
\name{print.cta_ort_summary}
\alias{print.cta_ort_summary}
\title{Print method for cta_ort_summary}
\usage{
\method{print}{cta_ort_summary}(x, ...)
}
\arguments{
\item{x}{A \code{cta_ort_summary} object.}

\item{...}{Unused.}
}
\value{
\code{invisible(x)}.
}
\description{
Print method for cta_ort_summary
}
