% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oda_s3.R
\name{print.oda_fit_summary}
\alias{print.oda_fit_summary}
\title{Print an ODA fit summary}
\usage{
\method{print}{oda_fit_summary}(x, ...)
}
\arguments{
\item{x}{An \code{oda_fit_summary} from \code{\link{summary.oda_fit}}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ODA fit summary
}
