% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_s3.R
\name{print.sda_fit}
\alias{print.sda_fit}
\title{Print an sda_fit object}
\usage{
\method{print}{sda_fit}(x, ...)
}
\arguments{
\item{x}{An \code{sda_fit} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}. Called primarily for its side effect of
  printing a human-readable summary of the SDA fit to the console.
}
\description{
Print an sda_fit object
}
