% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_s3.R
\name{print.sda_fit_summary}
\alias{print.sda_fit_summary}
\title{Print an sda_fit_summary object}
\usage{
\method{print}{sda_fit_summary}(x, ...)
}
\arguments{
\item{x}{An \code{sda_fit_summary} object.}

\item{...}{Unused.}
}
\description{
Print an sda_fit_summary object
}
