% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_s3.R
\name{summary.sda_fit}
\alias{summary.sda_fit}
\title{Summarise an sda_fit object}
\usage{
\method{summary}{sda_fit}(object, ...)
}
\arguments{
\item{object}{An \code{sda_fit} object.}

\item{...}{Unused.}
}
\description{
Summarise an sda_fit object
}
