% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta_family.R
\name{new_cta_family}
\alias{new_cta_family}
\title{Construct a CTA descendant family object (internal)}
\usage{
new_cta_family(members = list())
}
\arguments{
\item{members}{A list of objects from \code{\link{new_cta_family_member}}.}
}
\value{
A list of class \code{cta_family}.
}
\description{
Skeleton container for MDSA descendant family members built by
\code{\link{new_cta_family_member}}.  The full search loop
(\code{cta_descendant_family}) is deferred to Phase 2E.
}
\keyword{internal}
