% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cta_ort.R
\name{print.cta_ort}
\alias{print.cta_ort}
\title{Print method for Optimal Recursive Tree}
\usage{
\method{print}{cta_ort}(x, ...)
}
\arguments{
\item{x}{A \code{cta_ort} object.}

\item{...}{Unused.}
}
\value{
\code{invisible(x)}.
}
\description{
Print method for Optimal Recursive Tree
}
