% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_s3.R
\name{print.sda_fit}
\alias{print.sda_fit}
\title{Print an sda_fit object}
\usage{
\method{print}{sda_fit}(x, ...)
}
\arguments{
\item{x}{An \code{sda_fit} object.}

\item{...}{Unused.}
}
\description{
Print an sda_fit object
}
